intellij: spring-test

+++ imageserver-core/src/test/java/com/foreach/imageserver/core/data/ImageResolutionDaoTest.java
@@ -25,7 +25,7 @@ public class ImageResolutionDaoTest extends AbstractIntegrationTest {
 
         ImageResolution imageResolution = imageResolutionDao.getById(101010);
         assertNotNull(imageResolution);
-        assertEquals(101010, imageResolution.getId().intValue());
+        assertEquals(101010, imageResolution.getId());
         assertEquals(111, imageResolution.getWidth());
         assertEquals(222, imageResolution.getHeight());
     }
@@ -53,15 +53,15 @@ public class ImageResolutionDaoTest extends AbstractIntegrationTest {
         List<ImageResolution> imageResolutions = imageResolutionDao.getForContext(10);
         assertEquals(3, imageResolutions.size());
 
-        assertEquals(10, imageResolutions.get(0).getId().intValue());
+        assertEquals(10, imageResolutions.get(0).getId());
         assertEquals(111, imageResolutions.get(0).getWidth());
         assertEquals(222, imageResolutions.get(0).getHeight());
 
-        assertEquals(11, imageResolutions.get(1).getId().intValue());
+        assertEquals(11, imageResolutions.get(1).getId());
         assertNull(imageResolutions.get(1).getWidth());
         assertEquals(333, imageResolutions.get(1).getHeight());
 
-        assertEquals(12, imageResolutions.get(2).getId().intValue());
+        assertEquals(12, imageResolutions.get(2).getId());
         assertEquals(444, imageResolutions.get(2).getWidth());
         assertNull(imageResolutions.get(2).getHeight());
     }
